A partir desta alteração, uma nota só é removida do banco de dados caso tenha sido removida da API, o contrário disso fica com status de nota desativada. Alteração das funções de busca em NotaDao.kt para selecionar apenas as notas com status de desativada = 0. Alterado o retorno da função 'remove' em NotaWebClient, ao invés de utilizar o 'isSuccessful' passa-se a utilizar o 'true' para evitar problemas de remoção em notas que não existem na WebAPI.

+++ app/src/main/java/br/com/alura/ceep/repository/NotaRepository.kt
@@ -25,6 +25,11 @@ class NotaRepository(
     }
 
     suspend fun sincroniza() {
+        val notasDesativadas = dao.buscaDesativadas().first()
+        notasDesativadas.forEach { notaDesativada ->
+            remove(notaDesativada.id)
+
+        }
         val notasNaoSincronizadas = dao.buscaNaoSincronizadas().first()
         notasNaoSincronizadas.forEach { notaNaoSincronizada ->
             salva(notaNaoSincronizada)
@@ -37,8 +42,10 @@ class NotaRepository(
     }
 
     suspend fun remove(id: String) {
-        dao.remove(id)
-        webClient.remove(id)
+        dao.desativa(id)
+        if (webClient.remove(id)) {
+            dao.remove(id)
+        }
     }
 
     suspend fun salva(nota: Nota) {
